pyproject.toml
[tool.poetry]
name = "spectral"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
torch = "2.1.2"
torchvision = "^0.16.1"
ipykernel = "^6.25.2"
python = "3.11.7"
numpy = "^1.26.0"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
scipy = "^1.11.3"
lightning = "^2.1.2"
pyyaml = "^6.0.1"
scikit-image = "^0.24.0"
rich = "^13.7.1"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
